setup.py & packages
