+++ minionsai/discriminator_only/translator.py
@@ -1,8 +1,8 @@
 
 
 from typing import List
-from engine import Game, BOARD_SIZE
-from unit_type import unitList
+from ..engine import Game, BOARD_SIZE
+from ..unit_type import unitList
 import numpy as np
 
 class ObservationEnum():
