+++ test/run_game_test.py
@@ -1,0 +1,7 @@
+from minionsai.agent import CLIAgent, RandomAIAgent
+from minionsai.engine import Game
+from minionsai.run_game import run_game
+
+
+test_game = Game()
+run_game(test_game, (RandomAIAgent(), RandomAIAgent()), verbose=True)
